requirements.txt
requests
flask
flask-cors
gitpython
redis
rq